Switch level and subject IDs to Int64

+++ ios/AppStoreScreenshots/Screenshotter.swift
@@ -68,7 +68,7 @@ import WaniKaniAPI
   }
 
   class FakeLocalCachingClient: LocalCachingClient {
-    private let subjectTextProtos: [Int32: String] = [
+    private let subjectTextProtos: [Int64: String] = [
       // Subjects used in the search results for "sake". Abridged protos containing just the fields
       // displayed in search results.
       743: """
@@ -1262,7 +1262,7 @@ import WaniKaniAPI
       [86, 120, 485, 786, 2056]
     }
 
-    override func getSubject(id: Int32) -> TKMSubject? {
+    override func getSubject(id: Int64) -> TKMSubject? {
       if let textProto = subjectTextProtos[id] {
         return try! TKMSubject(textFormatString: textProto)
       }
@@ -1279,7 +1279,7 @@ import WaniKaniAPI
       return Array(repeating: a, count: availableReviewCount)
     }
 
-    override func getStudyMaterial(subjectId _: Int32) -> TKMStudyMaterials? {
+    override func getStudyMaterial(subjectId _: Int64) -> TKMStudyMaterials? {
       nil
     }
 
@@ -1295,7 +1295,7 @@ import WaniKaniAPI
       []
     }
 
-    override func getAssignment(subjectId _: Int32) -> TKMAssignment? {
+    override func getAssignment(subjectId _: Int64) -> TKMAssignment? {
       nil
     }
